requirements.txt

click>=6.0
pymongo>=3.8.0
PyYAML>=5.1
requests>=2.20.1
requests_oauthlib>=1.2.0
